cbb_model/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
jupyter>=1.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Phase 3: ML Model Dependencies
xgboost>=1.7.0
torch>=2.0.0
joblib>=1.3.0

# Phase 4: Model Optimization & Deployment
optuna>=3.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0